icache: line boundary: fix number of instructions
Each line in the icache would hold a certain number of instructions.
The regular instructions being inserted at positions 0 and 1 are both 4
byte instructions, which demand that we reduce the number of compressed
instructions by 1 to ensure that the target instruction has 2 bytes in
this line and the other 2 bytes elsewhere. Fix this by reducing the
range over which we iterate to generate the initial list of compressed
instructions.

Minor cleanup: avoid using '_' as a variable name. Fix this by
replacing '_' with list_iter inside the list comprehension.

Signed-off-by: Karthik B K <bkkarthik01@gmail.com>